Dự án python/shoe_store/requirements.txt
Django==4.2.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow>=11.0.0
django-filter==23.4
python-dotenv==1.0.0
django-cors-headers==4.3.1
